# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Problem": {
        "description": "An error as RFC 9457 describes it: the same shape whatever went wrong.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "what happened in this occurrence",
            "title": "Detail"
          },
          "instance": {
            "description": "path where it happened",
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "short human summary, the same for every occurrence",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "identifier of the kind of problem",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "instance"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "access token issued by the appliance Keycloak",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Campaigns, inventory, findings and evidence of the ARGOS appliance. Listings paginate by opaque cursor, errors follow RFC 9457 and every POST that creates accepts an Idempotency-Key.",
    "title": "ARGOS API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/findings/{finding_id}": {
      "get": {
        "operationId": "finding_api_v1_findings__finding_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Finding Api V1 Findings  Finding Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "no valid token"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the role does not allow it"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "it does not exist"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the state does not allow it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "invalid body"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "not configured yet"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "A finding with its whole why",
        "tags": [
          "findings"
        ]
      }
    },
...
  }
}